--- src/routes/components/table/docs/codeBlocks.ts
@@ -1,434 +1,0 @@
-export const simpleTableHTML = `
-<script lang="ts">
-	import { usersStore } from './data';
-</script>
-
-<div class="table-container">
-	<table class="table table-compact bg-tertiary-200">
-		<thead>
-			<tr class="bg-primary-300">
-				<th class="!p-2">Group</th>
-				<th class="!p-2">Description</th>
-			</tr>
-		</thead>
-		<tbody>
-			{#each $usersStore as user}
-				<tr>
-					<td>{user.name}</td>
-					<td>
-						<div>{user.group}</div>
-					</td>
-				</tr>
-			{/each}
-		</tbody>
-	</table>
-</div>`;
-
-export const groupHTML = `
-<script lang="ts">
-	import { Table } from '@bexis2/bexis2-core-ui';
-	import type { TableConfig } from '@bexis2/bexis2-core-ui';
-
-	import { groupsStore } from './data';
-	import type { Group } from './data';
-
-	const groupConfig: TableConfig<Group> = {						
-		id: 'userGroups',						
-		data: groupsStore,						
-		columns: {							
-			id: {								
-				header: 'ID',						
-			}						
-		}					
-	};
-</script>
-
-<Table config={groupConfig} />`;
-export const groupStoreCode = `			
-export type Group = {
-	id: number;
-	name: string;
-	description: string
-};
-
-const userGroups: Group[] = [					
-	{id: 1, name: 'Group 1', description: 'Testing 1'},					
-	{id: 2, name: 'Group 2', description: 'Testing 2'},					
-	{id: 3, name: 'Group 3', description: 'Testing 3'},					
-	...
-];					
-				
-export const groupsStore = writable<Group[]>(userGroups);`;
-
-export const usersHTML = `
-<script lang="ts">
-	import { Table } from '@bexis2/bexis2-core-ui';
-	import type { TableConfig } from '@bexis2/bexis2-core-ui';
-	
-	import { usersStore } from './data';
-	import type { User } from './data';
-
-	const usersConfig: TableConfig<User> = {
-		id: 'users',
-		data: usersStore
-	};
-</script>
-
-<Table config={usersConfig} />`;
-export const usersStoreCode = `
-export type User = { 
-	id: number; 
-	name: string; 
-	group: string; 
-	role: string 
-};
-
-const users: User[] = [
-	{id: 1, name: 'User 1', group: 'Group 1', role: 'Role 1'},
-	{id: 2, name: 'User 2', group: 'Group 2', role: 'Role 2'},
-	{id: 3, name: 'User 3', group: 'Group 3', role: 'Role 3'},
-	...
-];				
-				
-export const usersStore = writable<User[]>(users);`;
-export const usersNoRolesHTML = `
-<script lang="ts">
-	import { Table } from '@bexis2/bexis2-core-ui';
-	import type { TableConfig } from '@bexis2/bexis2-core-ui';
-
-	import TableFilter from './TableFilter.svelte';
-	import TableOptions from './TableOptions.svelte';
-	import { columnFilter } from './filters';
-	import { usersStore } from './data';
-	import type { User } from './data';
-
-	const usersNoRolesConfig: TableConfig<User> = {
-		id: 'usersNoRoles',
-		data: usersStore,
-		optionsComponent: TableOptions,
-		columns: {
-			id: {
-				colFilterFn: columnFilter,
-				colFilterComponent: TableFilter
-			},
-			role: {
-				exclude: true
-			},
-			group: {
-				header: 'Group name'
-			}
-		},
-		pageSizes: [1, 3, 5],
-		defaultPageSize: 5
-	};
-</script>
-
-<Table 
-	config={usersNoRolesConfig} 
-	on:action={(obj) => alert(obj.detail.type)} 
-/>`;
-
-export const usersBDHTML = `
-<script lang="ts">
-	import { Table } from '@bexis2/bexis2-core-ui';
-	import type { TableConfig } from '@bexis2/bexis2-core-ui';
-
-	import { usersBDStore } from './data';
-	import type { UserBD } from './data';
-
-	const usersBDConfig: TableConfig<UserBD> = {
-		id: 'usersBD',
-		data: usersBDStore,
-		columns: {
-			dateOfBirth: {
-				header: 'Date of birth',
-				instructions: {
-					toStringFn: 
-						(date: Date) =>
-							date.toLocaleString('en-US', { 
-								month: 'short', 
-								day: 'numeric', 
-								year: 'numeric' 
-							}
-						),
-					toSortableValueFn: 
-						(date: Date) => date.getTime(),
-					toFilterableValueFn: 
-						(date: Date) => date
-				}
-			}
-		}
-	};
-</script>
-
-<Table config={usersBDConfig} />`;
-
-export const usersBDStoreCode = `
-export type UserBD = { 
-	id: number; 
-	name: string; 
-	dateOfBirth: Date 
-};
-
-const usersBD = [
-	{
-		id: 1,
-		name: 'John Doe',
-		dateOfBirth: new Date('1990-05-10')
-	},
-	{
-		id: 2,
-		name: 'Jane Smith',
-		dateOfBirth: new Date('1985-12-15')
-	},
-	{
-		id: 3,
-		name: 'David Johnson',
-		dateOfBirth: new Date('1992-07-22')
-	},
-	...
-];
-
-export const usersBDStore = writable<UserBD[]>(usersBD);`;
-
-export const tableOptionsHTML = `
-<script lang="ts">
-	export let row;
-</script>
-
-<button 
-	class="btn btn-sm variant-filled-secondary" 
-	on:click|preventDefault={() => alert(JSON.stringify(row))}
->
-	See row info
-</button>
-`;
-
-export const tableConfigTypeCode = `
-export interface TableConfig<T> {
-	id: string;
-	data: Writable<T[]>;
-	columns?: Columns;
-	pageSizes?: number[];
-	defaultPageSize?: number;
-	optionsComponent?: typeof SvelteComponent;
-}`;
-
-export const columnsTypeCode = `
-export interface Columns {
-	[key: string]: Column;
-}`;
-
-export const columnTypeCode = `
-export interface Column {
-	header?: string;
-	exclude?: boolean; // false by default
-	instructions?: ColumnInstructions;
-	disableFiltering?: boolean; // false by default
-	disableSorting?: boolean; // false by default
-	colFilterFn?: ColumnFilterFn;
-	colFilterComponent?: typeof SvelteComponent;
-}`;
-
-export const columnInstructionsTypeCode = `
-export interface ColumnInstructions {
-	toStringFn?: (value: any) => string;
-	toSortableValueFn?: (value: any) => string | number;
-	toFilterableValueFn?: (value: any) => string | number | Date;
-}`;
-
-export const usersMissingIDsHTML = `
-<script lang="ts">
-	import { Table } from '@bexis2/bexis2-core-ui';
-	import type { TableConfig } from '@bexis2/bexis2-core-ui';
-
-	import { usersMissingIDsStore } from './data';
-	import type { UserMissingID } from './data';
-
-	const missingValues: { [key: number]: string } = {
-		32655: 'NA',
-		32654: 'NM',
-		32653: 'ND'
-	};
-
-	const usersMissingIDsConfig: TableConfig<UserMissingID> = {
-		id: 'usersMissingIDs',
-		data: usersMissingIDsStore,
-		columns: {
-			id: {
-				header: 'ID',
-				instructions: {
-					toStringFn: (key: number) => {
-						if (key in missingValues) {
-							return missingValues[key];
-						}
-						return key.toString();
-					}
-				}
-			}
-		}
-	};
-</script>
-
-<Table config={usersMissingIDsConfig} />`;
-
-export const usersMissingIDsStoreCode = `
-type UserMissingID = { 
-	id: number;
-	name: string; 
-	group: string; 
-	role: string 
-};
-	
-const usersMissingIDs = [
-	{
-		id: 1,
-		name: 'User 1',
-		group: 'Group 1',
-		role: 'Role 1'
-	},
-	{
-		id: 32655,
-		name: 'User 2',
-		group: 'Group 2',
-		role: 'Role 2'
-	},
-	{
-		id: 3,
-		name: 'User 3',
-		group: 'Group 3',
-		role: 'Role 3'
-	},
-	{
-		id: 32653,
-		name: 'User 4',
-		group: 'Group 4',
-		role: 'Role 4'
-	},
-	{
-		id: 5,
-		name: 'User 5',
-		group: 'Group 5',
-		role: 'Role 5'
-	}
-];
-
-export const usersMissingIDsStore = writable<UserMissingID[]>(usersMissingIDs);`;
-
-export const websitesHTML = `
-<script lang="ts">
-	import { Table } from '@bexis2/bexis2-core-ui';
-	import type { TableConfig } from '@bexis2/bexis2-core-ui';
-
-	import { websitesStore } from './data';
-	import type { Website } from './data';
-
-	type Website = { label: string; URL: URL };
-	const websitesStore = writable<Website[]>(websites);
-	const websitesConfig: TableConfig<Website> = {
-		id: 'websites',
-		data: websitesStore,
-		columns: {
-			URL: {
-				header: 'URL',
-				instructions: {
-					toStringFn: (url: URL) => url.toString(),
-					toFilterableValueFn: (url: URL) => url.toString()
-				},
-				disableSorting: true
-			}
-		}
-	};
-</script>
-
-<Table config={websitesConfig} />
-`;
-
-export const websitesStoreCode = `
-export type Website = { label: string; URL: URL };
-
-const websites: Website[] = [
-	{
-		label: 'Google',
-		URL: new URL('https://www.google.com')
-	},
-	{
-		label: 'GitHub',
-		URL: new URL('https://www.github.com')
-	},
-	{
-		label: 'YouTube',
-		URL: new URL('https://www.youtube.com')
-	},
-	{
-		label: 'Yahoo',
-		URL: new URL('https://www.yahoo.com')
-	},
-	{
-		label: 'Bing',
-		URL: new URL('https://www.bing.com')
-	}
-];
-
-export const websitesStore = writable<Website[]>(websites);
-`;
-
-export const usersAndAdminsHTML = `
-<script lang="ts">
-	import { Table } from '@bexis2/bexis2-core-ui';
-	import type { TableConfig } from '@bexis2/bexis2-core-ui';
-
-	import { usersAndAdminsStore } from './data';
-	import type { UserOrAdmin } from './data';
-
-	const usersAndAdminsConfig: TableConfig<UserOrAdmin> = {
-		id: 'usersAndAdmins',
-		data: usersAndAdminsStore,
-		columns: {
-			role: {
-				header: 'Admin',
-				instructions: {
-					toStringFn: (isAdmin: boolean) => (isAdmin ? '✓' : '')
-				},
-				disableFiltering: true
-			}
-		}
-	};
-</script>`;
-export const usersAndAdminsStoreCode = `
-export type UserOrAdmin = {
-	name: string;
-	username: string;
-	isAdmin: boolean;
-};
-
-const usersAndAdmins: UserOrAdmin[] = [
-	{
-		name: 'John Doe',
-		username: 'johndoe',
-		isAdmin: true
-	},
-	{
-		name: 'Jane Smith',
-		username: 'janesmith',
-		isAdmin: false
-	},
-	{
-		name: 'Robert Johnson',
-		username: 'robertjohnson',
-		isAdmin: false
-	},
-	{
-		name: 'Emily Davis',
-		username: 'emilydavis',
-		isAdmin: true
-	},
-	{
-		name: 'Michael Wilson',
-		username: 'michaelwilson',
-		isAdmin: false
-	}
-];
-
-export const usersAndAdminsStore = writable<UserOrAdmin[]>(usersAndAdmins);
-`;
